Add Node tests for angle helpers
