+++ index.html
@@ -30,6 +30,7 @@
         </div>
     </div>
 
+    <script src="angles.js"></script>
     <script src="index.js"></script>
 </body>
 </html>
